requirements.txt
# Web Scraping
beautifulsoup4==4.11.1
selenium==4.7.0
requests==2.28.0
playwright==1.35.0

# Data Processing
pandas==1.5.0
numpy==1.23.0

# API Integrations
anthropic==0.8.0
google-api-python-client==2.70.0
pydrive==1.3.1

# Utilities
python-dotenv==1.0.0
retry==0.9.2

# Development Tools
pytest==7.3.1
black==23.3.0
pylint==2.17.4
jupyter==1.0.0

# Task Queues and Scheduling (for future implementation)
celery==5.2.7
redis==4.5.5
airflow==2.6.1

# ML/AI Libraries (for future implementation)
scikit-learn==1.2.2
transformers==4.28.1
spacy==3.5.3

# FastAPI (for microservices)
fastapi==0.95.2
uvicorn==0.22.0

# MongoDB Driver
pymongo==4.3.3